refactor: move document onSubmit to its own hook
