+++ packages/outstatic/src/pages/edit-document/index.tsx
@@ -3,7 +3,7 @@ import matter from 'gray-matter'
 import Head from 'next/head'
 import { useRouter } from 'next/router'
 import { singular } from 'pluralize'
-import { useContext, useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import showdown from 'showdown'
 import {
@@ -12,54 +12,32 @@ import {
   DocumentSettings,
   DocumentTitleInput
 } from '../../components'
-import { OutstaticContext, DocumentContext } from '../../context'
+import { DocumentContext } from '../../context'
 import { useCreateCommitMutation } from '../../graphql/generated'
-import {
-  CustomFieldArrayValue,
-  CustomFields,
-  Document,
-  FileType,
-  isArrayCustomField
-} from '../../types'
+import { CustomFields, Document, FileType } from '../../types'
 import { useOstSession } from '../../utils/auth/hooks'
 import { IMAGES_PATH } from '../../utils/constants'
 import { deepReplace } from '../../utils/deepReplace'
 import { escapeRegExp } from '../../utils/escapeRegExp'
 import { getLocalDate } from '../../utils/getLocalDate'
-import { mergeMdMeta } from '../../utils/mergeMdMeta'
 import { replaceImageSrcRoot } from '../../utils/replaceImageSrc'
 import useNavigationLock from '../../utils/useNavigationLock'
-import useOid from '../../utils/useOid'
 import useTipTap from '../../utils/useTipTap'
 import { convertSchemaToYup, editDocumentSchema } from '../../utils/yup'
-import { createCommit as createCommitApi } from '../../utils/createCommit'
-import { assertUnreachable } from '../../utils/assertUnreachable'
-import { MetadataSchema } from '../../utils/metadata/types'
-import { hashFromUrl } from '../../utils/hashFromUrl'
-import MurmurHash3 from 'imurmurhash'
-import { stringifyMetadata } from '../../utils/metadata/stringify'
 import useFileQuery from '../../utils/useFileQuery'
+import useSubmitDocument from '../../utils/hooks/useSubmitDocument'
 
-type EditDocumentProps = {
-  collection: string
-}
-
-export default function EditDocument({ collection }: EditDocumentProps) {
+export default function EditDocument({ collection }: { collection: string }) {
   const router = useRouter()
   const [slug, setSlug] = useState(router.query?.ost?.[1] as string)
-  const { repoOwner, repoSlug, repoBranch, contentPath, monorepoPath } =
-    useContext(OutstaticContext)
   const { session } = useOstSession()
   const [loading, setLoading] = useState(false)
   const [hasChanges, setHasChanges] = useState(false)
   const [files, setFiles] = useState<FileType[]>([])
   const [createCommit] = useCreateCommitMutation()
-  const fetchOid = useOid()
   const [showDelete, setShowDelete] = useState(false)
   const [documentSchema, setDocumentSchema] = useState(editDocumentSchema)
-  const methods = useForm<Document>({
-    resolver: yupResolver(documentSchema)
-  })
+  const methods = useForm<Document>({ resolver: yupResolver(documentSchema) })
   const { editor } = useTipTap({ ...methods })
   const [customFields, setCustomFields] = useState<CustomFields>({})
 
@@ -78,187 +56,20 @@ export default function EditDocument({ collection }: EditDocumentProps) {
     file: `${collection}/schema.json`
   })
 
-  const { data: metadata } = useFileQuery({
-    file: `metadata.json`
+  const onSubmit = useSubmitDocument({
+    session,
+    slug,
+    setSlug,
+    setShowDelete,
+    setLoading,
+    files,
+    createCommit,
+    methods,
+    collection,
+    customFields,
+    setCustomFields,
+    setHasChanges
   })
-
-  const onSubmit = async (data: Document) => {
-    setLoading(true)
-    try {
-      const document = methods.getValues()
-      let content = mergeMdMeta({ ...data })
-      const oid = await fetchOid()
-      const owner = repoOwner || session?.user?.login || ''
-      const newSlug = document.slug
-
-      // If the slug has changed, commit should delete old file
-      const oldSlug = slug !== newSlug && slug !== 'new' ? slug : undefined
-
-      const capi = createCommitApi({
-        message: oldSlug
-          ? `chore: Updates ${newSlug} formerly ${oldSlug}`
-          : `chore: Updates/Creates ${newSlug}`,
-        owner,
-        oid: oid ?? '',
-        name: repoSlug,
-        branch: repoBranch
-      })
-
-      if (oldSlug) {
-        capi.removeFile(
-          `${
-            monorepoPath ? monorepoPath + '/' : ''
-          }${contentPath}/${collection}/${oldSlug}.md`
-        )
-      }
-
-      if (files.length > 0) {
-        files.forEach(({ filename, blob, type, content: fileContents }) => {
-          // check if blob is still in the document before adding file to the commit
-          if (blob && content.search(blob) !== -1) {
-            const randString = window
-              .btoa(Math.random().toString())
-              .substring(10, 6)
-            const newFilename = filename
-              .toLowerCase()
-              .replace(/[^a-zA-Z0-9-_\.]/g, '-')
-              .replace(/(\.[^\.]*)?$/, `-${randString}$1`)
-
-            const filePath = (() => {
-              switch (type) {
-                case 'images':
-                  return IMAGES_PATH
-                default:
-                  assertUnreachable(type)
-              }
-            })()
-
-            capi.replaceFile(
-              `${
-                monorepoPath ? monorepoPath + '/' : ''
-              }public/${filePath}${newFilename}`,
-              fileContents,
-              false
-            )
-
-            // replace blob in content with path
-            content = content.replace(blob, `/${filePath}${newFilename}`)
-          }
-        })
-      }
-
-      const { data: matterData } = matter(content)
-
-      capi.replaceFile(
-        `${
-          monorepoPath ? monorepoPath + '/' : ''
-        }${contentPath}/${collection}/${newSlug}.md`,
-        content
-      )
-
-      // Check if a new tag value was added
-      let hasNewTag = false
-      Object.entries(customFields).forEach(([key, field]) => {
-        const customField = customFields[key]
-
-        // Only check for new values in array fields
-        if (isArrayCustomField(field) && isArrayCustomField(customField)) {
-          // @ts-ignore
-          data[key].forEach((selectedTag: CustomFieldArrayValue) => {
-            // Check if the selected tag already exists
-            const exists = field.values.some(
-              (savedTag: CustomFieldArrayValue) =>
-                savedTag.value === selectedTag.value
-            )
-
-            // If the selected tag does not exist, add it
-            if (!exists) {
-              customField.values.push({
-                value: selectedTag.value,
-                label: selectedTag.label
-              })
-              customFields[key] = customField
-              setCustomFields({ ...customFields })
-              hasNewTag = true
-            }
-          })
-        }
-      })
-
-      if (hasNewTag) {
-        const customFieldsJSON = JSON.stringify(
-          {
-            title: collection,
-            type: 'object',
-            properties: { ...customFields }
-          },
-          null,
-          2
-        )
-
-        capi.replaceFile(
-          `${
-            monorepoPath ? monorepoPath + '/' : ''
-          }${contentPath}/${collection}/schema.json`,
-          customFieldsJSON + '\n'
-        )
-      }
-
-      // update metadata for this post
-      // requires final content for hashing
-      if (metadata?.repository?.object?.__typename === 'Blob') {
-        const m = JSON.parse(
-          metadata.repository.object.text ?? '{}'
-        ) as MetadataSchema
-        m.generated = new Date().toISOString()
-        m.commit = hashFromUrl(metadata.repository.object.commitUrl)
-        const newMeta = (m.metadata ?? []).filter(
-          (c) =>
-            !(
-              c.collection === collection &&
-              (c.slug === oldSlug || c.slug === newSlug)
-            )
-        )
-        const state = MurmurHash3(content)
-        newMeta.push({
-          ...matterData,
-          title: matterData.title,
-          publishedAt: matterData.publishedAt,
-          status: matterData.status,
-          slug: newSlug,
-          collection,
-          __outstatic: {
-            hash: `${state.result()}`,
-            commit: m.commit,
-            path: `${contentPath}/${collection}/${newSlug}.md`
-          }
-        })
-
-        capi.replaceFile(
-          `${
-            monorepoPath ? monorepoPath + '/' : ''
-          }${contentPath}/metadata.json`,
-          stringifyMetadata({ ...m, metadata: newMeta })
-        )
-      }
-
-      const input = capi.createInput()
-
-      await createCommit({
-        variables: {
-          input
-        }
-      })
-      setLoading(false)
-      setHasChanges(false)
-      setSlug(newSlug)
-      setShowDelete(true)
-    } catch (error) {
-      // TODO: Better error treatment
-      setLoading(false)
-      console.log({ error })
-    }
-  }
 
   useEffect(() => {
     router.push(`/outstatic/${collection}/${slug}`, undefined, {
